Normalize sample library size to effective size

This is a helper function for handle_input. edgeR::calcNormFactors function is used to estimate normalizing factors, which is used to multiply library sizes.

effective_size(outlist, outRle, genome = "hg19", nc = 2, verbose = FALSE)

Arguments

outlist

a list of list objects with four elements, 'query' is either a GRanges object or an RleList object, 'size' is the library size, 'type' is the input file type, 'weight' is the name of the metadata column

outRle

logical, indicating whether the 'query' element of the output should be an RleList object or a GRanges object

genome

a string denoting the genome name and version

nc

integer, number of cores for parallel processing

verbose

logical, whether to output additional information

Value

a list of list objects with four elements ('query', 'size', 'type', 'weight'), with the 'size' element modified.

Author

Shuye Pu

Examples

queryFiles <- system.file("extdata", "chip_treat_chr19.bam", package = "GenomicPlot")
names(queryFiles) <- "query"

inputFiles <- system.file("extdata", "chip_input_chr19.bam", package = "GenomicPlot")
names(inputFiles) <- "input"

importParams <- list(
  offset = 0, fix_width = 140, fix_point = "start", norm = FALSE,
  useScore = FALSE, outRle = FALSE, useSizeFactor = FALSE, genome = "hg19"
)

out_list <- handle_input(
  inputFiles = c(queryFiles, inputFiles),
  importParams = importParams, verbose = TRUE, nc = 2
)
#> Reading bamfile: C:/Users/greenblatt/AppData/Local/Temp/RtmpiQE4hr/temp_libpath167c637a5c73/GenomicPlot/extdata/chip_treat_chr19.bam
#> Paired end bam file detected!
#> 
#> Bam file C:/Users/greenblatt/AppData/Local/Temp/RtmpiQE4hr/temp_libpath167c637a5c73/GenomicPlot/extdata/chip_treat_chr19.bam is loaded
#> Cached .rds file is modified using new input parameters
#> Reading bamfile: C:/Users/greenblatt/AppData/Local/Temp/RtmpiQE4hr/temp_libpath167c637a5c73/GenomicPlot/extdata/chip_input_chr19.bam
#> Paired end bam file detected!
#> 
#> Bam file C:/Users/greenblatt/AppData/Local/Temp/RtmpiQE4hr/temp_libpath167c637a5c73/GenomicPlot/extdata/chip_input_chr19.bam is loaded
#> Cached .rds file is modified using new input parameters

res <- effective_size(out_list, outRle = TRUE, genome = "hg19", nc = 2, verbose = TRUE)
#> Estimating size factor
#> Library size before adjusting: 114163117433
#> Library size after adjusting: 104841127873
#> Library size normalizing factors: 0.9183493950709971.08891017445783