Filter GRanges by overlaps in a stranded way

This function reports all query GRanges that have overlaps in subject GRanges. Strand information is used to define overlap.

filter_by_overlaps_stranded(query, subject, maxgap = -1L, minoverlap = 0L)

Arguments

query

a GRanges object

subject

a GRanges object

maxgap

an integer denoting the distance that define overlap

minoverlap

The minimum amount of overlap between intervals as a single integer greater than 0. If you modify this argument, maxgap must be held fixed.

Value

a GRanges object

Author

Shuye Pu

Examples

query <- GRanges("chr19", 
  IRanges(rep(c(10, 15), 2), width=c(1, 20, 40, 50)), 
  strand=c("+", "+", "-", "-")
)

subject <- GRanges("chr19", 
  IRanges(rep(c(13, 150), 2), width=c(10, 14, 20, 28)), 
  strand=c("+", "-", "-", "+")
)

res <- filter_by_overlaps_stranded(query, subject)
res
#> GRanges object with 3 ranges and 0 metadata columns:
#>       seqnames    ranges strand
#>          <Rle> <IRanges>  <Rle>
#>   [1]    chr19     15-34      +
#>   [2]    chr19     10-49      -
#>   [3]    chr19     15-64      -
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths