Get genomic coordinates of features of protein-coding genes

Get genomic coordinates of promoter, 5'UTR, CDS, 3'UTR, TTS and intron for the longest transcript of protein-coding genes. The range of promoter is defined by fiveP and dsTSS upstream and downstream TSS, respectively, the TTS ranges from the 3' end of the gene to threeP downstream, or the start of a downstream gene, whichever is closer.

get_txdb_features(txdb, fiveP = -1000, dsTSS = 300, threeP = 1000, nc = 2)

Arguments

txdb

a TxDb object defined in the GenomicFeatures package

fiveP

extension upstream of the 5' boundary of genes

dsTSS

range of promoter extending downstream of TSS

threeP

extension downstream of the 3' boundary of genes

nc

number of cores for parallel processing

Value

a GRangesList object

Author

Shuye Pu

Examples

txdb <- AnnotationDbi::loadDb(system.file("extdata", "txdb_chr19.sql", 
package="GenomicPlot"))

f <- get_txdb_features(txdb, dsTSS=100, fiveP=-100, threeP=100)