Handle files in bw|bigwig|bigWig|BigWig|BW|BIGWIG format

This is a function for read NGS coverage data in bigwig format, store the input data in a list of GRanges objects or RleList objects. The input bw file can be stranded or non-stranded. Library size is calculate as the sum of all coverage.

handle_bw(inputFile, importParams, verbose = FALSE)

Arguments

inputFile

a string denoting path to the input file

importParams

a list of parameters, refer to handle_input for details

verbose

logical, whether to output additional information

Value

a list object with four elements, 'query' is a list GRanges objects or RleList objects, 'size' is the estimated library size, 'type' is the input file type, weight' is the name of the metadata column to be used as weight for coverage calculation

Details

For stranded files, forward and reverse strands are stored in separate files, with '+' or 'p' in the forward strand file name and '-' or 'm' in the reverse strand file name.

Author

Shuye Pu

Examples

queryFiles <- system.file("extdata", "test_wig_chr19_+.bw", package = "GenomicPlot")
names(queryFiles) <- "test_bw"

importParams <- list(
  offset = 0, fix_width = 0, fix_point = "start", norm = FALSE,
  useScore = FALSE, outRle = TRUE, useSizeFactor = FALSE, genome = "hg19"
)

out <- handle_bw(queryFiles, importParams, verbose = TRUE)
#> Mate found:
#> C:/Users/greenblatt/AppData/Local/Temp/RtmpiQE4hr/temp_libpath167c637a5c73/GenomicPlot/extdata/test_wig_chr19_+.bw
#> C:/Users/greenblatt/AppData/Local/Temp/RtmpiQE4hr/temp_libpath167c637a5c73/GenomicPlot/extdata/test_wig_chr19_-.bw
#> -
lapply(out$query, sum)
#> $chr1
#> [1] 0
#> 
#> $chr2
#> [1] 0
#> 
#> $chr3
#> [1] 0
#> 
#> $chr4
#> [1] 0
#> 
#> $chr5
#> [1] 0
#> 
#> $chr6
#> [1] 0
#> 
#> $chr7
#> [1] 0
#> 
#> $chr8
#> [1] 0
#> 
#> $chr9
#> [1] 0
#> 
#> $chr10
#> [1] 0
#> 
#> $chr11
#> [1] 0
#> 
#> $chr12
#> [1] 0
#> 
#> $chr13
#> [1] 0
#> 
#> $chr14
#> [1] 0
#> 
#> $chr15
#> [1] 0
#> 
#> $chr16
#> [1] 0
#> 
#> $chr17
#> [1] 0
#> 
#> $chr18
#> [1] 0
#> 
#> $chr19
#> [1] 1180957
#> 
#> $chr20
#> [1] 0
#> 
#> $chr21
#> [1] 0
#> 
#> $chr22
#> [1] 0
#> 
#> $chrX
#> [1] 0
#> 
#> $chrY
#> [1] 0
#> 
#> $chrM
#> [1] 0
#>