Make TxDb object from a GTF file for a subset of genes

Make a partial TxDb object given a GTF file and a list of gene names in a file or in a character vector.

make_subTxDb_from_GTF(gtfFile, geneList, geneCol = 1)

Arguments

gtfFile

path to a GTF file

geneList

path to a tab-delimited text file with one gene name on each line, or a character vector of gene names

geneCol

the position of the column that containing gene names in the case that geneList is a file

Value

a TxDb object

Author

Shuye Pu

Examples

gtfFile <- system.file("extdata", "gencode.v19.annotation_chr19.gtf", package = "GenomicPlot") 
genes <- c("RPRD1A", "RPAP2", "RPRD1B", "RPRD2", "ZNF281", "ZNF121", "YTHDF2")

txdb <- make_subTxDb_from_GTF(gtfFile = gtfFile, geneList = genes)
#> Reading existing granges.rds object from C:/Users/greenblatt/AppData/Local/Temp/RtmpiQE4hr/temp_libpath167c637a5c73/GenomicPlot/extdata/gencode.v19.annotation_chr19.gtf.granges.rds
#> Keeping standard chromosomes only
#> File C:/Users/greenblatt/AppData/Local/Temp/RtmpiQE4hr/temp_libpath167c637a5c73/GenomicPlot/extdata/gencode.v19.annotation_chr19.gtf.granges.rds already exists.
#>                   Use overwriteObjectAsRds = TRUE to overwrite the file
#> Warning: The "phase" metadata column contains non-NA values for features of type
#>   stop_codon. This information was ignored.