Preprocess scoreMatrix before plotting

This is a helper function for manipulate the score matrix produced by ScoreMatrix or ScoreMatrinBin functions defined in the 'genomation' package. To facilitate downstream analysis, imputation of missing values is performed implicitly.

process_scoreMatrix(
  fullmatrix,
  scale = FALSE,
  rmOutlier = 0,
  transform = NA,
  verbose = FALSE
)

Arguments

fullmatrix

a numeric matrix, with bins in columns and genomic windows in rows

scale

logical, indicating whether the score matrix should be scaled to the range 0:1, so that samples with different baseline can be compared

rmOutlier

logical, indicating whether a row with abnormally high values in the score matrix should be removed

transform

a string in c("log", "log2", "log10"), default = NA indicating no transformation of data matrix

verbose

logical, indicating whether to output additional information (data used for plotting or statistical test results)

Value

a numeric matrix with the same dimension as the fullmatrix

Details

If inputFiles for the plotting function is null, all operations (impute, scale, rmOutlier and transform) can be applied to the score matrix, in the order of impute -> rmOutlier -> transform -> scale. When inputFiles are provided, only impute and rmOutlier can be applied to the score matrix, as transform and scale will affect ratio calculation, especially when log2 transformation of the ratio is intended. However, all these operations can be applied to the resulting ratio matrix. In order to avoid introducing distortion into the post-processed data, use caution when applying these operations.

Author

Shuye Pu

Examples

fullMatrix <- matrix(rlnorm(100), ncol = 10)
for (i in 5:6) {
  fullMatrix[i, 4:7] <- NaN
  fullMatrix[i+1, 4:7] <- NA
  fullMatrix[i+2, 4:7] <- -Inf
  fullMatrix[i-1, 4:7] <- 0
  fullMatrix[i-2, 1:3] <- Inf
}
fullMatrix[9, 4:7] <- runif(4) + 90

wo <- process_scoreMatrix(fullMatrix, rmOutlier = 3, verbose = TRUE)
#> Inspecting matrix
#>          n_invalid fraction_invalid
#> infinite        10             0.10
#> NA               8             0.08
#> NaN              4             0.04
#> zero             8             0.08
#> Imputing missing values...
#> Matrix quartiles:
#> 0 0 0.687252945778437 2.03539311347142 90.9118664462585
#> 
#> Matrix quantiles after imputing:
#> 0.0345654936779461 0.0345654936779461 0.687252945778437 2.03539311347142 90.9118664462585
#> The imputed value is: 0.0345654936779461
#> Outlier detected:
#> Maximum of the matrix 90.9118664462585
#> 
#> Median of row max 4.79206042859457
#> 
#> Median absolute deviation (mad) of row max 3.40096177135035
#> 
#> Mulitplier of mad 3
#> 
#> Up_bound and replace value 14.9949457426456
#> 
#> percentile of up_bound0.95
#> 
#> Pumber of outlier rows2
#> 
#> Number of outliers5
#> 
#> Fraction of outliers 0.05
#> 
#> Values of outliers:
#> 90.2051386407111 90.9118664462585 90.7047755708918 90.4543539362494 32.083498725835
tf <- process_scoreMatrix(fullMatrix, 
  rmOutlier = 0, transform = "log2", verbose = TRUE
)
#> Inspecting matrix
#>          n_invalid fraction_invalid
#> infinite        10             0.10
#> NA               8             0.08
#> NaN              4             0.04
#> zero             8             0.08
#> Imputing missing values...
#> Matrix quartiles:
#> 0 0 0.687252945778437 2.03539311347142 90.9118664462585
#> 
#> Matrix quantiles after imputing:
#> 0.0345654936779461 0.0345654936779461 0.687252945778437 2.03539311347142 90.9118664462585
#> The imputed value is: 0.0345654936779461
scaled <- process_scoreMatrix(fullMatrix, scale = TRUE, verbose = TRUE)
#> Inspecting matrix
#>          n_invalid fraction_invalid
#> infinite        10             0.10
#> NA               8             0.08
#> NaN              4             0.04
#> zero             8             0.08
#> Imputing missing values...
#> Matrix quartiles:
#> 0 0 0.687252945778437 2.03539311347142 90.9118664462585
#> 
#> Matrix quantiles after imputing:
#> 0.0345654936779461 0.0345654936779461 0.687252945778437 2.03539311347142 90.9118664462585
#> The imputed value is: 0.0345654936779461